requirements.txt
streamlit==1.31.0
psycopg2-binary==2.9.9
pandas==2.1.4
rapidfuzz==3.6.1
requests==2.31.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
beautifulsoup4
requests
